test/com/twu/biblioteca/WelcomeMessageTest.java
package com.twu.biblioteca;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import static org.junit.Assert.assertEquals;

public class WelcomeMessageTest {

    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
    PrintStream defaultOutStream = System.out;

    @Before
    public void setUpStreams() {
        System.setOut(new PrintStream(outContent));
    }

    @After
    public void cleanUpStreams() {
        System.setOut(defaultOutStream);
    }
    @Test
    public void shouldDisplayWelcomeMessage() {
        WelcomeMessage message = new WelcomeMessage();

        message.show();

        assertEquals("\t!!..Welcome to the Bibliotica..!!\n", outContent.toString());
    }
}
